move dashobard component to index page
